New question tab and upvote now working.

+++ src/store/steemqa/actions.js
@@ -1,18 +1,8 @@
 import axios from 'axios'
 
-export const topics = (state) => {
+export const favouriteTopicsById = (state, payload) => {
   return axios.get(
-    state.getters.serverURL + '/topic/'
-  ).then(function (response) {
-    state.commit('topics', response.data)
-  }).catch(function (error) {
-    console.log(error)
-  })
-}
-
-export const favouriteTopics = (state, payload) => {
-  return axios.get(
-    state.getters.serverURL + '/favourite_topic/',
+    state.getters.serverURL + '/favourite_topics/',
     {
       params: {
         username: payload.username,
@@ -20,7 +10,7 @@ export const favouriteTopics = (state, payload) => {
       }
     }
   ).then(function (response) {
-    state.commit('favouriteTopics', response.data)
+    state.commit('favouriteTopicsById', response.data)
   }).catch(function (error) {
     console.log(error)
   })
@@ -28,32 +18,35 @@ export const favouriteTopics = (state, payload) => {
 
 export const addTopic = (state, payload) => {
   return axios.post(
-    state.getters.serverURL + '/favourite_topic/',
+    state.getters.serverURL + '/favourite_topics/',
     {
       username: payload.username,
       access_token: payload.accessToken,
       topic: payload.id
     }
   ).then(function (response) {
-    state.commit('addFavouriteTopic', response.data.id)
+    state.commit('addFavouriteTopic', response.data)
   }).catch(function (error) {
     console.log(error)
   })
 }
 
 export const removeTopic = (state, payload) => {
-  return axios.delete(
-    state.getters.serverURL + '/favourite_topic/',
-    {
-      params: {
-        username: payload.username,
-        access_token: payload.accessToken,
-        topic: payload.id
-      }
+  for (let topic of state.getters.favouriteTopicsById) {
+    if (topic[1].topic.toString() === payload.id) {
+      return axios.delete(
+        state.getters.serverURL + '/favourite_topics/' + topic[0] + '/',
+        {
+          params: {
+            username: payload.username,
+            access_token: payload.accessToken
+          }
+        }
+      ).then(function (response) {
+        state.commit('removeFavouriteTopic', topic[0])
+      }).catch(function (error) {
+        console.log(error)
+      })
     }
-  ).then(function (response) {
-    state.commit('removeFavouriteTopic', payload.id)
-  }).catch(function (error) {
-    console.log(error)
-  })
+  }
 }
